Complete on full functionality in interactive.js
